+++ internal/provider/provider.go
@@ -358,7 +358,7 @@ func New(version, userAgent string) func() *schema.Provider {
 				"confluent_schema_registry_cluster_mode":       schemaRegistryClusterModeDataSource(),
 				"confluent_user":                               userDataSource(),
 				"confluent_invitation":                         invitationDataSource(),
-				"confluent_byok_key":                           byokDataSource(),
+				"confluent_byok_key":                           byokKeyDataSource(),
 				"confluent_network_link_endpoint":              networkLinkEndpointDataSource(),
 				"confluent_network_link_service":               networkLinkServiceDataSource(),
 				"confluent_tableflow_topic":                    tableflowTopicDataSource(),
@@ -374,7 +374,7 @@ func New(version, userAgent string) func() *schema.Provider {
 			ResourcesMap: map[string]*schema.Resource{
 				"confluent_catalog_integration":                catalogIntegrationResource(),
 				"confluent_api_key":                            apiKeyResource(),
-				"confluent_byok_key":                           byokResource(),
+				"confluent_byok_key":                           byokKeyResource(),
 				"confluent_certificate_authority":              certificateAuthorityResource(),
 				"confluent_certificate_pool":                   certificatePoolResource(),
 				"confluent_cluster_link":                       clusterLinkResource(),